pyproject.toml
[project]
name = "neuro-trail"
version = "0.1.0"
description = "Personalized Memory Augmented Learning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ebooklib>=0.18",
    "neo4j>=5.28.1",
    "dotenv>=0.9.9",
    "streamlit>=1.42.2",
    "litellm>=1.61.20",
    "ruff>=0.9.7",
    "tenacity>=9.0.0",
    "numpy==1.26.4",
    "langchain-community>=0.3.18",
    "chromadb>=0.6.3",
    "googleapis-common-protos>=1.68.0",
    "faiss-cpu>=1.10.0",
]

[project.packages]
find = { where = ["src"] }

[tool.uv.workspace]
members = ["venv"]

[tool.ruff]
lint.select = ["E", "F", "I", "TID251"]
lint.ignore = ["E501"]
line-length = 88
indent-width = 4
extend-include = ["*.ipynb"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
